+++ src/index.ts
@@ -27,6 +27,7 @@ export * from "./widgets/select-field/select-field";
 export * from "./widgets/multi-select-field/multi-select-field";
 
 export * from "./utils/use-query-params/use-query-params";
+export * from "./utils/use-countdown/use-countdown";
 
 export * from "../src-old/form/file-picker";
 export * from "../src-old/form/file-field";
@@ -35,7 +36,6 @@ export * from "../src-old/form/date-field";
 export * from "../src-old/form/time-picker";
 export * from "../src-old/form/time-field";
 
-export * from "../src-old/hooks/use-countdown";
 export * from "../src-old/hooks/use-load-more";
 export * from "../src-old/hooks/use-window-breakpoints";